修复、验证博客后台管理增删改，优化使用体验。

+++ blogserver/src/main/java/cn/xmh/web/blogserver/model/DaysData.java
@@ -22,5 +22,5 @@ public class DaysData implements Serializable {
     private int views;
     private int likes;
     private int comments;
-    private int articleId;
+    private long articleId;
 }
